recommend-parking-function/requirements.txt
azure-functions
azure-storage
googlemaps
jsonschema
faker